root/Backend_folder/requirements.txt
alembic==1.13.2
amqp==5.3.1
aniso8601==9.0.1
backports.functools-lru-cache==2.0.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.8.2
Brotli==1.1.0
cachelib==0.9.0
celery==5.4.0
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-Excel==0.0.7
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.55.0
greenlet==3.0.3
httplib2==0.22.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
lml==0.1.0
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
numpy==2.1.3
pandas==2.2.3
pdf_reports==0.3.7
pillow==10.4.0
prompt_toolkit==3.0.48
pycparser==2.22
pydyf==0.11.0
pyexcel==0.7.1
pyexcel-io==0.6.7
pyexcel-webio==0.1.4
PyJWT==2.9.0
pyparsing==3.2.0
pyphen==0.15.0
pypugjs==5.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.2.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.32
texttable==1.7.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.0
webencodings==0.5.1
Werkzeug==3.0.4
WTForms==3.1.2
zopfli==0.2.3.post1
